requirements.txt
itsdangerous==0.24
dnspython==1.15.0
Flask==0.9
Flask-Admin==1.5.0
Flask-Mail==0.9.1
Flask-PyMongo==0.5.1
GeoAlchemy2==0.4.0
geocoder==1.20.1
GeoIP==1.3.2
geopy==1.11.0
gunicorn==19.4.5
ipaddress==1.0.17
ipdetailscache==0.4.8
Jinja2==2.9.6
Localization==0.1.4
MarkupSafe==1.0
MongoAlchemy==0.21
nmap==0.0.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
pydns==2.3.6
pygeoip==0.3.2
pymongo==3.4.0
python-dateutil==2.5.3
python-whois==0.6.5
pywhois==0.1
requests==2.13.0
SQLAlchemy==1.1.11
sqlalchemy-mixins==0.1.4
Werkzeug==0.12.2
whois==0.7
ipwhois==1.0.0
urllib3==1.21.1
folium==0.3.0
